requirements.txt
# Core ML/DS stack
mlflow==2.12.1
pandas>=1.5.3
scikit-learn>=1.3.0
numpy
scipy>=1.10.0
imblearn

# Google BigQuery support
google-cloud-bigquery>=3.12.0
google-auth>=2.20.0

# Optional: fast model serving
fastapi==0.111.0
uvicorn==0.29.0

# Optional: serialize with pickle
cloudpickle>=2.2.1

optuna>=3.1.0
xgboost
seaborn